Update legacy-ediscovery-retirement.md
Can we add an additional statement regarding Hybrid deployment to the Note.  We are getting questions on whether this applies if cloud mailboxes are in the same on-prem eDiscovery search.  Just need the statement to clarify this.  It is documented here: https://docs.microsoft.com/en-us/exchange/security-and-compliance/in-place-ediscovery/in-place-ediscovery